[Passed] 실수 입력받아 계산 (48ms, 16MB)

+++ trail1/실수 입력받아 계산/calculation-by-inputting-a-real-number.py
@@ -1,4 +1,4 @@
-a = int(input())
-b = int(input())
+a = float(input())
+b = float(input())
 
 print(f"{a + b:.2f}")
